Improve if-env directive, does not render element under wrong env
The if-env directive was modified to a behaviour closer to ng-if
The entire element is transcluded and if the ENV matches it appends
it self back to the DOM.

Priority chosen for this directive is the same as ng-if (600).
It also lacks the watch functionallity of ng-if as there is no reason
the ENV variable will change on runtime.